# Optimize RTen implementation