… whisper/whisper-rs/Cargo.lock
… whisper/whisper-rs/Cargo.toml