Thanks again for putting this demo together. It has been very useful for finding optimizations for transformer models.

This PR contains a series of optimizations to the whisper-rs demo. On my Intel i5 laptop this reduces transcription time for the 2 min test audio from ~28s to ~8.75s (see [CHANGELOG](https://github.com/robertknight/rten/blob/main/CHANGELOG.md))

- Update to rten v0.7.0. This includes various optimizations that benefit transformers and autoregressive models in particular
- Reduce tensor copies in this repository as much as possible. In particular, avoid copying data when converting ndarrays to rten tensor views for model inputs, and when converting rten tensors to ndarray arrays for model outputs
- Use the new `Model::partial_run` API to precompute parts of the decoder graph that do not change in each iteration of the loop. It turns out that the most expensive part of the decoder model is computing the cross-attention keys and values. This depends only on the encoder outputs, and so computing them once per audio chunk, instead of at each decoder timestep, saves a lot of time

Of the above optimizations, the first two reduce runtime from ~28s to ~20s. The precomputation with `Model::partial_run` accounts for the rest of the improvement.

The result is now much more competitive with whisper.cpp on my system, which takes ~9s with the default beam size of 5 and ~6.8s with a beam size of 1 (which I think is equivalent to the greedy decoding that this code does).